backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
PyMongo==4.6.0
bcrypt==4.1.2
python-dotenv==1.0.0
marshmallow==3.20.1
email-validator==2.1.0
